Categorizer/requirements.txt
requests
jsonpickle
pymongo
pika
mysql-connector-python
pymongo[srv]